+++ src/main/java/pe/edu/lamolina/pivot/controller/rolexamen/gruporegular/GrupoRegularServiceImp.java
@@ -439,9 +439,9 @@ public class GrupoRegularServiceImp implements GrupoRegularService {
 
         List<SeccionGrupoRegular> seccionesGruposRegularesExc = seccionGrupoRegularDAO.allByRolExamenes(rolExamenes, SeccionRolExamenEstadoEnum.EXC);
         List<Seccion> seccionesExcluidas = seccionesGruposRegularesExc.stream().map(x -> x.getSeccion()).collect(Collectors.toList());
-
-        seccionExcluidoDAO.deleteBySecciones(seccionesExcluidas);
-
+        if (!seccionesExcluidas.isEmpty()) {
+            seccionExcluidoDAO.deleteBySecciones(seccionesExcluidas);
+        }
         for (LetraGrupoRegular letraGrupoRegular : letrasGruposRegular) {
             alumnoGrupoRegularDAO.deleteByLetraGrupoRegular(letraGrupoRegular);
             seccionGrupoRegularDAO.deleteByLetraGrupoRegular(letraGrupoRegular);
